[Implementation], [Testing], and [UI]
Implemented comparable protocol and < method in Restaurant so Restaurant List's Restaurant Array can be sorted. Used so active restaurants are always at the top of the home screen.
Added new tests and updated tests to test comparable and sorting functionality.
Updated UI so unavailable restaurants can still be click in navigation view, but button and details are not available.
Added GPX file so simulator can pretend to be near campus.